fix: remove circular imports for server functions and queries
