+++ packages/app-builder/src/components/ClientDetail/AddConfigurationModal.tsx
@@ -1,9 +1,7 @@
 import { useLoaderRevalidator } from '@app-builder/contexts/LoaderRevalidatorContext';
 import { DataModel } from '@app-builder/models';
-import {
-  addConfigurationPayloadSchema,
-  useAddConfigurationMutation,
-} from '@app-builder/queries/client360/add-configuration';
+import { useAddConfigurationMutation } from '@app-builder/queries/client360/add-configuration';
+import { addConfigurationPayloadSchema } from '@app-builder/schemas/client360';
 import { handleSubmit } from '@app-builder/utils/form';
 import { useForm, useStore } from '@tanstack/react-form';
 import { Client360Table } from 'marble-api';
