+++ packages/app-builder/src/constants/client360.ts
@@ -1,0 +1,1 @@
+export const semanticTypes = ['person', 'company'] as const;
