+++ packages/app-builder/src/queries/client360/add-configuration.ts
@@ -1,18 +1,7 @@
+import { AddConfigurationPayload } from '@app-builder/schemas/client360';
 import { addClient360ConfigurationFn } from '@app-builder/server-fns/client-360';
 import { useMutation } from '@tanstack/react-query';
 import { useServerFn } from '@tanstack/react-start';
-import { z } from 'zod/v4';
-
-const semanticTypes = ['person', 'company'] as const;
-
-export const addConfigurationPayloadSchema = z.object({
-  tableId: z.uuid(),
-  semanticType: z.enum(semanticTypes),
-  captionField: z.string().min(1),
-  alias: z.string().optional(),
-});
-
-export type AddConfigurationPayload = z.infer<typeof addConfigurationPayloadSchema>;
 
 export const useAddConfigurationMutation = () => {
   const addConfiguration = useServerFn(addClient360ConfigurationFn);
