+++ packages/app-builder/src/queries/scoring/update-settings.ts
@@ -1,13 +1,7 @@
+import { type UpdateScoringSettingsPayload } from '@app-builder/schemas/user-scoring';
 import { updateScoringSettingsFn } from '@app-builder/server-fns/scoring';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { useServerFn } from '@tanstack/react-start';
-import { z } from 'zod/v4';
-
-export const updateScoringSettingsPayloadSchema = z.object({
-  maxRiskLevel: z.number(),
-});
-
-export type UpdateScoringSettingsPayload = z.infer<typeof updateScoringSettingsPayloadSchema>;
 
 export const useUpdateScoringSettingsMutation = () => {
   const updateScoringSettings = useServerFn(updateScoringSettingsFn);
